Add lint job in the github action workflow (#6)
* github/action: add actions-rs toolchain setup

Use actions-rs to setup the toolchain which will be used by lint steps.

* github/action: add rustfmt check

Assert that style is respect for all merge request before the costly
linting and compilation pipeline is started. No merge request can be
incorrectly accepted after this change.

* github/action: add clippy lint

Deny any warnings found by clippy before compiling the project, avoiding
merge requests introducing new ones without tagging. It is meant to help
iterating over a merge request with feedbacks but only merge it when it
is ready.

* github/action: enable rust workflow for merge requests

It is now time to enable the full testsuite for new merge requests and
begin a mandatory review workflow.

* core/enums: remove duplicate comment

The comment was unaligned and lead to error with rustfmt --check.
Removing it in favor of the one from the documentation above which tells
the same thing.

* rustfmt: split stable features from unstable

* core/image: allow dead code on `content`

The member is currently unused and prevent cargo clippy from being
happy. It is ought to be removed someday when the `content` member will
get used.

* README: add github action badge

The badge is autogenerated from github action service and increase the
confidence in the crate for newcomers. It's also very fun.